TPMA_backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.5.7
# cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.1.0
django-extensions==3.2.3
django-smtp-ssl==1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
oauthlib==3.2.2
packaging==23.1
# Pillow==9.5.0
protobuf==4.23.3
# psycopg2==2.9.6
# psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
